refactor: refactor: add courseid routes to admin base route and remove components
BREAKING CHANGE: every component that was in pages directory is removed and every page created in
admin page that is
