+++ package.json
@@ -9,7 +9,8 @@
     "lint": "next lint",
     "prepare": "husky install",
     "commit": "git-cz",
-    "test": "jest"
+    "test": "jest",
+    "tsc": "tsc"
   },
   "dependencies": {
     "@heroicons/react": "^2.0.12",
