+++ src/components/admin/pageBase/index.tsx
