+++ src/pages/api/auth/[...nextauth].ts
@@ -1,4 +1,4 @@
-import NextAuth, { type NextAuthOptions } from 'next-auth'
+import NextAuth, { NextAuthOptions } from 'next-auth'
 
 export const authOptions: NextAuthOptions = {
   // Include user.id on session
@@ -8,10 +8,10 @@ export const authOptions: NextAuthOptions = {
         session.user.id = user.id
       }
       return session
-    }
+    },
   },
   // Configure one or more authentication providers
-  providers: []
+  providers: [],
 }
 
 export default NextAuth(authOptions)
